requirements.txt
attrs==24.2.0
cattrs==24.1.2
certifi==2024.8.30
cfbd==4.5.2
charset-normalizer==3.4.0
exceptiongroup==1.2.2
flatbuffers==24.3.25
idna==3.10
numpy==2.0.2
openmeteo_requests==1.3.0
openmeteo_sdk==1.18.0
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
plotly==5.24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
six==1.16.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
url-normalize==1.4.3
urllib3==2.2.3
